backend/backend/requirements/base.txt
Django==3.0.5
pytest-django==3.9.0
python-dotenv==0.17.1
psycopg2-binary==2.8.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.16.0
pydantic==1.8.1
django-filter==2.2.0
django-cors-headers==3.2.1
django-extensions==2.2.9
social-auth-app-django==3.1.0
slackclient==2.5.0
ipdb
django-extensions==2.2.9
factory-boy==3.2.0

# fix install issue
cryptography==3.3.2

# Gunicorn
gunicorn==20.0.4
gevent==1.4.0

# swagger
drf-yasg[validation]==1.20.0
